Fix emacsql/buit-in sqlite incompatibility
Emacsql correctly was storing integers as integers, whereas the builtin sqlite
was both reading and writing integers as strings.  This corrects the built-in
sqlite to use integers instead of strings.

Fixes the cause of https://github.com/ahyatt/triples/issues/2, although we will
need a further upgrading mechanism for data written in the built-in sqlite.