Fix more fundamental differences between sqlite and emacsql
Previously, for unknown reasons (probably just a mistake), the built-in sqlite
version of the subject column was TEXT, while the emacsql had no
type (defaulting it to BLOB).  This fixes that discrepancy, so now the sqlite
subject column also is created with no type.

Additionally, the previous version fixing the data could lose data, since it
would transform any strings starting with a number, while it should just
transform strings that are complete numbers.

Because of these changes, we now just always move the existing triples to a temp
table, create a new one (the right way now), copy everything over, then go
through the subjects one by one to make sure we only change things that are
really numbers.

This does not handle any cases where users create a database with sqlite, but
transition to 0.3 using emacsql, which we assume will not happen.  At any rate,
from 0.3 on, things will hopefully be correct by default.